Split off into bare_root_metadata.go
